requirements.txt
asgiref==3.5.2
Django==4.1.3
django-crum==0.7.9
django-environ==0.9.0
django-widget-tweaks==1.4.12
fontawesomefree==6.2.0
mysqlclient==2.1.1
Pillow==9.3.0
PyMySQL==1.0.2
sqlparse==0.4.3
tzdata==2022.6
